requirements.txt
fastapi
uvicorn
trafilatura
google-genai
pydantic